Trying to setup heketi deployment
The problem is, that the glusterfs container has a big
problem. It internally uses lvm: this is currently not usable
because of a bug with a timeout and udev:
WARNING: Device /dev/vdb not initialized in udev database even after waiting 10000000 microseconds.

Need - also for a limited time - to support host based glusterfs.

Signed-off-by: Andreas Florath <andreas.florath@telekom.de>